Make code more reusable in other contexts - Export processStream methods, they are useful when using the code outside of Istio - Move verifySentResourcesMultipleTypes to client_test.go
