+++ pkg/mcp/internal/size.go
@@ -1,0 +1,11 @@
+// Provides a ProtoSize method that can be swapped for proto.Size when using Google protos.
+
+package internal
+
+type hasSize interface {
+	Size() int
+}
+
+func ProtoSize(msg hasSize) int {
+	return msg.Size()
+}
